requirements.txt
streamlit
tensorflow-cpu
pydicom
numpy



